requirements.txt
aiosmtplib==1.1.6
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
cryptography==36.0.2
Deprecated==1.2.13
ecdsa==0.18.0
fastapi==0.79.0
fastapi-jwt-auth==0.5.0
filelock==3.7.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
jsonschema==3.2.0
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
phonenumbers==8.12.48
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.10.4
pydantic==1.9.1
PyJWT==2.3.0
pyrsistent==0.18.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
supertokens-python==0.10.2
tldextract==3.1.0
twilio==7.9.1
typing_extensions==4.1.1
urllib3==1.26.11
uvicorn==0.18.2
Werkzeug==2.0.3
wrapt==1.14.1
zipp==3.8.1
